+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - The Wound Threshold optional rule was no longer working
 - Dexterity bonuses to CMD were incorrectly limited by the actor's Maximum Dexterity Bonus
 - The slayer talent class feature was incorrectly gained at level 1
+- Some attributes (hp, max xp, and some others) were incorrect after importing an actor
 
 ### Changelog